Moving some tests utilities to tests/utils.rs
